57585 - Add redis decriptions deleteValidations & remove updateDescriptionLinkValidations
