+++ src/server/cache/repository/validation/description/index.ts
@@ -1,9 +1,9 @@
+import { deleteValidations } from 'server/cache/repository/validation/description/deleteValidations'
 import { getValidations } from 'server/cache/repository/validation/description/getValidations'
 import { setValidations } from 'server/cache/repository/validation/description/setValidations'
-import { updateDescriptionLinkValidations } from 'server/cache/repository/validation/description/updateDescriptionLinkValidations'
 
 export const DescriptionValidationRedisRepository = {
+  deleteValidations,
   getValidations,
   setValidations,
-  updateDescriptionLinkValidations,
 }
